Use nuxt server init
